fit a GLM with l0, l1, or tlp regularization

Fit a generalize linear model via penalized maximum likelihood. The regularization path is computed for the l0, lasso, or truncated lasso penalty at a grid of values for the regularization parameter num of nonzeros or lambda. Can deal with all shapes of data (not large sparse data matrices now). Fits linear, logistic, multinomial, and poisson models.

The sequence of models implied by lambda or kappa is fit by coordinate descent.

The package provides 3 penalties: l0, l1, and tlp and 3 distribution families: gaussian, binomial, and poisson.

glmtlp(
  X,
  y,
  family = c("gaussian", "binomial"),
  penalty = c("l0", "l1", "tlp"),
  nlambda = 100,
  lambda.min.ratio = ifelse(nobs < nvars, 0.01, 1e-04),
  lambda = NULL,
  kappa = NULL,
  tau = 0.3 * sqrt(log(nvars)/nobs),
  delta = 2,
  tol = 1e-04,
  weights = NULL,
  penalty.factor = rep(1, nvars),
  standardize = TRUE,
  dc.maxit = 20,
  cd.maxit = 10000,
  nr.maxit = 500,
  ...
)

Arguments

X
input matrix, of dimension nobs x nvars; each row is an observation vector. Currently, doesn't support the sparse matrix format (inherit from class "sparseMatrix" as in package Matrix).

y
response variable, of length nobs. Quantitative for family="gaussian", or family="poisson" (non-negative counts). For family="binomial", should be either a factor with two levels. For family="multinomial", (to be added).

family
a character string representing one of the built-in families. See Details section below.

penalty
a character string representing one of the built-in penalties. "l0" represents L_0 penalty, "l1" represents lasso-type penalty, and "tlp" represents truncated lasso penalty.

nlambda
The number of lambda values. Default is 100.

lambda.min.ratio
Smallest value for lambda, as a fraction of lambda.max, the (data derived) entry value (i.e. the smallest value for which all coefficients are zero). The default depends on the sample size nobs relative to the number of variables nvars. If nobs > nvars, the default is 0.0001, close to zero. If nobs < nvars, the default is 0.01.

lambda
A user supplied lambda sequence. Typical usage is to have the program compute its own lambda sequence based on nlambda and lambda.min.ratio. Supplying a value of lambda overrides this. WARNING: use with care. Avoid supplying a single value for lambda (for predictions after CV use predict() instead). Supply instead a decreasing sequence of lambda values. glmtlp relies on its warms starts for speed, and its often faster to fit a whole path than compute a single fit. When the penalty is 'l0', it is not recommended for the users to supply this parameter.

kappa
A user supplied kappa sequence. Typical usage is to have the program compute its own kappa sequence based on nvars and nobs. This sequence is used when penalty is 'l0'.

tau
A tuning parameter used in the TLP penalty. Default is 0.3 * sqrt(log(nvars)/nobs).

delta
A tuning parameter used in the coordinate majorization descent algorithm. See Yang, Y., & Zou, H. (2014) in the reference for detail.

tol
Tolerance level for all the iterative optimization algorithms.

weights
observation weights. Default is 1 for each observation

penalty.factor
Separate penalty factors can be applied to each coefficient. This is a number that multiplies lambda to allow differential shrinkage. Can be 0 for some variables, which implies no shrinkage, and that variable is always included in the model. Default is 1 for all variables.

standardize
Logical. Whether or not standardize the input matrix X; default is TRUE.

dc.maxit
Maximum number of iterations for the DC (Difference of Convex Functions) programming; default is 20.

cd.maxit
Maximum number of iterations for the coordinate descent algorithm; default is 10^4.

nr.maxit
Maximum number of iterations for the Newton-Rapson method; default is 500.

...
Additional argument. These include some of the original arguments to 'glmtlp', and each must be named if used.

Value

An object with S3 class "glmtlp".

beta

a nvars x length(kappa) matrix of coefficients when penalty is 'l0'; a nvars x length(lambda) matrix of coefficients when penalty is 'l1' and 'tlp'.

call

the call that produced this object.

family

the distribution family used in the model fitting.

intercept

the intercept vector, of length(kappa) or length(lambda).

lambda

The actual sequence of lambda values used. Note that the length may be smaller than the provided nlambda due to removal of saturated values.

penalty

the penalty type in the model fitting.

penalty.factor

the penalty factor for each coefficient used in the model fitting.

tau

the tuning parameter used in the model fitting, available when penalty = 'tlp'.

user.lambda

logical, whether or not the used lambda sequence is provided by the user.

user.kappa

logical, whether or not the used kappa sequence is provided by the user, available when penalty = 'l0'.

Details

The objective function for "gaussian" is $$1/2 RSS/nobs + \lambda*penalty,$$ and for the other models it is $$-loglik/nobs + \lambda*penalty.$$ Note also that for "gaussian", glmtlp standardizes y to have unit variance (using 1/(n-1) formula). The coefficients for any predictor variables with zero variance are set to zero for all values of lambda.

## Details on `family` option

glmtlp currently only supports built-in families.

The built in families are specified via a character string. For all families, the object produced is a regularization path for fitting the generalized linear regression paths, by maximizing the appropriate penalized log-likelihood. Sometimes the sequence is truncated before nlambda values of lambda have been used, because of instabilities in the inverse link functions near a saturated fit. glmtlp(..., family="binomial") fits a traditional logistic regression model for the log-odds. glmtlp(..., family="multinomial") fits a symmetric multinomial model, where each class is represented by a linear model (on the log-scale). The penalties take care of redundancies. A two-class "multinomial" model will produce the same fit as the corresponding "binomial" model, except the pair of coefficient matrices will be equal in magnitude and opposite in sign, and half the "binomial" values.

## Details on `penalty` option

The built in penalties are specified by a character string. For l0 penalty, kappa sequence is used for generating the regularization path, while for the other penalties, lambda sequence is used for generating the regularization path.

glmtlp functions

`glmtlp()`, `cv.glmtlp()`

References

Shen, X., Pan, W., & Zhu, Y. (2012). Likelihood-based selection and sharp parameter estimation. Journal of the American Statistical Association, 107(497), 223-232.
Yang, Y., & Zou, H. (2014). A coordinate majorization descent algorithm for l1 penalized learning. Journal of Statistical Computation and Simulation, 84(1), 84-95.
Two R packge Github: ncvreg and glmnet.

See also

Author

Chunlin Li, Yu Yang
Maintainer: Yu Yang yang6367@umn.edu

Examples

# Gaussian
X <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
fit1 <- glmtlp(X, y, family = "gaussian", penalty = "l0")
fit2 <- glmtlp(X, y, family = "gaussian", penalty = "l1")
fit3 <- glmtlp(X, y, family = "gaussian", penalty = "tlp")

# Binomial

X <- matrix(rnorm(100 * 20), 100, 20)
y <- sample(c(0,1), 100, replace = TRUE)
fit <- glmtlp(X, y, family = "binomial", penalty = "l1")