make predictions from a "cv.glmtlp" object.

This function makes predictions from a cross-validated glmtlp model, using the stored "glmtlp" object, and the optimal value chosen for lambda.

# S3 method for cv.glmtlp
predict(
  object,
  X,
  type = c("link", "response", "class", "coefficients", "numnzs", "varnzs"),
  lambda = NULL,
  kappa = NULL,
  which = object$idx.min,
  ...
)

# S3 method for cv.glmtlp
coef(object, lambda = NULL, kappa = NULL, which = object$idx.min, ...)

Arguments

object
Fitted "cv.glmtlp" object.

X
Matrix of new values for x at which predictions are to be made. Must be a matrix. See documentation for predict.glmtlp.

type
Type of prediction required. Type "link" gives the linear predictors for "binomial" models; for "gaussian" models it gives the fitted values. Type "response" gives the fitted probabilities for "binomial"; for "gaussian" type "response" is equivalent to type "link". Type "coefficients" computes the coefficients at the requested values for lambda or kappa. Note that for "binomial" models, results are returned only for the class corresponding to the second level of the factor response. Type "class" applies only to "binomial" models, and produces the class label corresponding to the maximum probability. Type "numnz" returns the total number of nonzeros coefficients for each value of lambda or kappa. Type "varnz" returns a list of the indices of the nonzero coefficients for each value of lambda or kappa.

lambda
Value of the penalty parameter lambda at which predictions are required. Default is NULL.

kappa
Value of the penalty parameter kappa at which predictions are required. Default is NULL.

which
Index of the penalty parameter lambda or kappa sequence at which predictions are required. Default is the idx.min stored on the CV object.

...
Additional arguments.

Value

The object returned depends on the ... argument which is passed on to the predict method for glmtlp objects.

Details

This function makes it easier to use the results of cross-validation to make a prediction.

References

Shen, X., Pan, W., & Zhu, Y. (2012). Likelihood-based selection and sharp parameter estimation. Journal of the American Statistical Association, 107(497), 223-232.
Yang, Y., & Zou, H. (2014). A coordinate majorization descent algorithm for l1 penalized learning. Journal of Statistical Computation and Simulation, 84(1), 84-95.
Two R packge Github: ncvreg and glmnet.

See also

Author

Chunlin Li, Yu Yang
Maintainer: Yu Yang yang6367@umn.edu

Examples

X <- matrix(rnorm(100 * 20), 100, 20)
y <- rnorm(100)
cv.fit <- cv.glmtlp(X, y, family = "gaussian", penalty = "l1", ncores=2)
predict(cv.fit, X = X[1:5, ])

#> [1] -0.0385279  0.2072894  0.1714689  0.1171646  0.1203098

coef(cv.fit)

#>   intercept          V1          V2          V3          V4          V5 
#>  0.08881710  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000 
#>          V6          V7          V8          V9         V10         V11 
#>  0.00000000  0.00000000  0.00000000  0.00000000 -0.14000879  0.00000000 
#>         V12         V13         V14         V15         V16         V17 
#>  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000 -0.03037846 
#>         V18         V19         V20 
#>  0.00000000  0.00000000  0.00000000 

predict(cv.fit, X = X[1:5, ], lambda = 0.1)

#> [1] -0.1825810  0.2703292  0.2246965  0.1082700  0.1361068